#212 - Clean up imports, rename backend support file
