+++ src/utils/src/index.ts
@@ -8,7 +8,7 @@ export * from './types/project-types';
 export * from './types/user-types';
 export * from './types/api-utils-types';
 
-export * from './backend-supports/all-projects';
+export * from './backend-supports/projects-get-all';
 
 export * from './dummy-data';
 export * from './validate-wbs';
